Create IoT Based Forest Fire Warning System using Node mcu
Here we gonna predict the fire by observing the changes in the environmental factors like temperature, humidity, light intensity and so on using the Iot technology with the microcontroller Nodemcu and few sensors (ultraSonic , temperature and humidity, soil moisture sensor). We have used only 3 sensors but many other sensors can be added if we want a more accurate analysis.